# lib/api-schemas/reco-proba-post-body.ts
import { z } from 'zod';

/**
 * Shared POST body for `/api/recommendation` and `/api/probability` (D.62).
 * Unknown top-level keys are allowed for forward compatibility.
 */
export const recoProbaPostBodySchema = z
  .object({
    profile: z.record(z.string(), z.unknown()).optional(),
    playground: z.boolean().optional(),
    /** Optional: pin this country first in ranked results (public numeric id only). */
    focusCountryId: z.number().int().positive().optional(),
  })
  .passthrough();

export type RecoProbaPostBody = z.infer<typeof recoProbaPostBodySchema>;